pyproject.toml
[project]
authors = [{ name = "pablo vela", email = "pablovela5620@gmail.com" }]
name = "vistadream"
requires-python = ">= 3.11"
version = "0.1.0"
dependencies = [
    "gradio",
    "gradio-rerun>=0.24,<=0.27",
    "rerun-notebook>=0.24.0,<=0.27",
    "einops>=0.8.1,<0.9",
    # 允许在设置了 SOCKS 代理(例如 ALL_PROXY=socks5://...)的环境中正常 import gradio/httpx
    "socksio>=1.0.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["https://prefix.dev/conda-forge"]
platforms = ["linux-64"]

[tool.pixi.system-requirements]
cuda = "12"

[tool.pixi.activation]
# You can find your CUDA GPU Compute Capability using the command `nvidia-smi --query-gpu=compute_cap --format=csv` or here: https://developer.nvidia.com/cuda-gpus 
env = { TORCH_CUDA_ARCH_LIST = "7.5 8.6 8.9 12.0 12.1" }

[tool.pixi.dependencies]
python = "3.12.*"
cmake = ">=4.0.3,<5"
# CUDA Build Tools
# You can find your CUDA version using the command `nvidia-smi` or `nvcc --version`
cuda-version = "12.9.*"
cuda-compiler = "*"
cuda-cudart-dev = "*"
cuda-crt = "*"
libcusparse-dev = "*"
cuda-driver-dev = "*"
cuda-nvcc = "*"
cuda-nvrtc-dev = "*"
cuda-nvtx = "*"
cuda-nvtx-dev = "*"
cuda-nvml-dev = "*"
cuda-profiler-api = "*"
# CUDA Libraries
cudnn = "*"
libcublas-dev = "*"
libcudss-dev = "*"
libcufile-dev = "*"
libcufft-dev = "*"
libcurand-dev = "*"
libcusolver-dev = "*"
cusparselt = "*"
libnvjitlink = "*"
# end of CUDA dependencies
scikit-learn = ">=1.7.1,<2"
jupyterlab = ">=4.4.4,<5"
open3d = ">=0.19.0,<0.20"
nodejs = ">=24.4.1,<24.5"
pytorch-gpu = "*"
torchvision = "*"
torchmetrics = "*"


[tool.pixi.pypi-options]
no-build-isolation = ["gsplat", "xformers"]

[tool.pixi.pypi-dependencies]
vistadream = { path = ".", editable = true }
gsplat = { git = "https://github.com/nerfstudio-project/gsplat.git", rev = "2323de5905d5e90e035f792fe65bad0fedd413e7" }
accelerate = "*"
tqdm = "*"
plyfile = "*"
timm = "*"
diffusers = "*"
imageio = "*"
imageio-ffmpeg = "*"
transformers = "*"
torchsde = "*"
huggingface-hub = "*"
roma = "*"
peft = "*"
omegaconf = "*"
jaxtyping = ">=0.2.36,<0.3"

# xformers = { git = "https://github.com/facebookresearch/xformers.git", rev = "main" }
simplecv = { git = "https://github.com/pablovela5620/simplecv.git", rev = "a75e2769bfab5d098df8d0e17112f6242f73d4ee" }
monopriors = { git = "https://github.com/pablovela5620/monoprior.git" }
moge = { git = "https://github.com/microsoft/MoGe.git", rev = "6b8b43db567ca4b08615c39b42cffd6c76cada29" }
vggt = { git = "https://github.com/pablovela5620/vggt.git", rev = "5ed6ec88a951f27959f73bc72eb4b6dc7a028b0c" }

[tool.pixi.feature.dev.dependencies]
beartype = ">=0.19.0,<0.20"
ruff = ">=0.12.7,<0.13"
pytest = ">=8.4.1,<9"

[tool.pixi.feature.dev.pypi-dependencies]
ipython_beartype = { git = "https://github.com/beartype/ipython-beartype.git", rev = "main" }

[tool.pixi.environments]
dev = { features = ["dev"] }

[tool.pixi.tasks]
example = { cmd = "python tools/run_single_img.py --image-path data/office/IMG_4029.jpg", depends-on = [
    "_install-splat-transform",
    "_download-models",
] }

[tool.pixi.tasks._download-models]
cmd = """
    test -e ckpt/vec.pt
    || (
        huggingface-cli download pablovela5620/vistadream  --repo-type model \
        --local-dir ckpt/
    )
    """
outputs = ["ckpt/vec.pt"]
description = "Downloads models from huggingface"

[tool.pixi.tasks._install-splat-transform]
cmd = """
    test -e $CONDA_PREFIX/bin/splat-transform
    || (
        npm install -g @playcanvas/splat-transform
    )
    """
outputs = ["$CONDA_PREFIX/bin/splat-transform"]
description = "Installs the splat-transform tool"

[tool.ruff]
line-length = 150
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

ignore = [
    "E501",  # Line too long.
    "F722",  # Forward annotation false positive from jaxtyping. Should be caught by pyright.
    "F821",  # Forward annotation false positive from jaxtyping. Should be caught by pyright.
    "UP037", # Quoted annotations, false positive from jaxtyping.
]
